Finish importing mixin to all interceptors
